+++ memgpt/server/rest_api/routers/v1/__init__.py
@@ -1,5 +1,7 @@
 from memgpt.server.rest_api.routers.v1.tools import tools_router
+from memgpt.server.rest_api.routers.v1.sources import sources_router
 
 ROUTERS = [
     tools_router,
+    sources_router,
 ]
